Move PostgreSQL utilities into new pkg
Some of these functions could be potentially useful for other projects
and have a logical separation from the code they were currently mixed
with. This also makes some of the code a bit more manageable to look
through.

In particular, parsing pg_config's version output may be useful in
other tools, so it'd be nice to have it available in one package
that can be linked against.
