+++ internal/parse/parse.go
@@ -119,29 +119,3 @@ func PGFormatToBytes(val string) (uint64, error) {
 	}
 	return ret, nil
 }
-
-// ToPGMajorVersion returns the major PostgreSQL version associated with a given
-// version string, as given from an invocation of pg_config --version. This string
-// has the form of "PostgreSQL X.Y[.Z (extra)]". For versions before 10, the major
-// version is defined as X.Y, whereas starting with 10 it is defined as just X.
-// That is, "PostgreSQL 10.3" returns "10" and "PostgreSQL 9.6.4" returns "9.6".
-func ToPGMajorVersion(val string) (string, error) {
-	res := pgVersionRegex.FindStringSubmatch(val)
-	if len(res) != 3 {
-		return "", fmt.Errorf(errCouldNotParseVersionFmt, val)
-	}
-	switch res[1] {
-	case "11":
-		fallthrough
-	case "10":
-		return res[1], nil
-	case "9":
-		fallthrough
-	case "8":
-		fallthrough
-	case "7":
-		return res[1] + "." + res[2], nil
-	default:
-		return "", fmt.Errorf(errUnknownMajorVersionFmt, val)
-	}
-}
